Switch to cleaner approach for testing trycp integration
